🧹 CLEANUP: Repository cleaned for v3.0
REMOVED old files:
- ❌ Old copier.py (2101 lines) → ✅ telegram_copier_v3.py (845 lines)
- ❌ Old main.py → ✅ main.py (v3.0 implementation)
- ❌ All test files (test_*.py) - old tests for broken v2.x
- ❌ All debugging/patch files - no longer needed
- ❌ All old documentation (ALBUM_*, CRITICAL_*, etc.)
- ❌ Old state files (copied_messages.json, last_message_id.txt)

KEPT essential files:
- ✅ telegram_copier_v3.py - New core implementation
- ✅ main.py - Clean entry point
- ✅ config.py - Configuration management
- ✅ utils.py - Shared utilities
- ✅ requirements.txt - Dependencies
- ✅ README.md - Updated for v3.0
- ✅ UPGRADE_TO_V3_GUIDE.md - Migration guide
- ✅ USAGE.md - Quick start guide
- ✅ LICENSE - License file
- ✅ .gitignore - Git settings

Repository is now clean and ready for v3.0 production use!
All critical issues are fixed in the new implementation.
