--- ALBUM_FIX_COMMENTS.md
@@ -1,153 +1,0 @@
-# Исправление критической проблемы с альбомами в комментариях
-
-## 🚨 Описание проблемы
-
-При копировании альбомов из комментариев (discussion groups) скрипт работал некорректно:
-- ✅ Текст комментариев копировался правильно
-- ✅ Хронологический порядок сохранялся  
-- ❌ **Фотографии из альбомов прикреплялись как "unnamed" файлы без расширения**
-- ❌ При скачивании этих файлов они оказывались битыми (требовали ручного добавления `.jpg`)
-
-## 🔍 Причина проблемы
-
-Проблема находилась в методах `copy_album()` и `copy_single_message()` в файле `copier.py`.
-
-### До исправления:
-```python
-# Строки 907-908 в copy_album()
-downloaded_file = await self.client.download_media(message.media, file=bytes)
-media_files.append(downloaded_file)
-```
-
-При скачивании медиа как `bytes` объекты, **Telethon терял всю информацию о файле**:
-- ❌ Имя файла
-- ❌ Расширение файла  
-- ❌ MIME-тип
-- ❌ Атрибуты документа
-
-При отправке через `send_file()`, Telegram не знал, как правильно отобразить файл → сохранял как "unnamed" без расширения.
-
-## ✅ Решение
-
-Исправление заключается в **сохранении атрибутов файлов при скачивании** и передаче их при отправке.
-
-### После исправления:
-
-#### 1. Извлечение информации о файле
-```python
-# Получаем информацию о файле из оригинального медиа
-original_attributes = None
-original_mime_type = None
-suggested_filename = None
-
-if hasattr(message.media, 'document') and message.media.document:
-    doc = message.media.document
-    original_attributes = doc.attributes if hasattr(doc, 'attributes') else []
-    original_mime_type = getattr(doc, 'mime_type', None)
-    
-    # Пытаемся извлечь имя файла из атрибутов
-    from telethon.tl.types import DocumentAttributeFilename
-    for attr in original_attributes:
-        if isinstance(attr, DocumentAttributeFilename):
-            suggested_filename = attr.file_name
-            break
-```
-
-#### 2. Генерация имени файла на основе MIME-типа
-```python
-# Если имя файла не найдено, генерируем на основе MIME-типа
-if not suggested_filename:
-    if original_mime_type:
-        if original_mime_type.startswith('image/'):
-            extension = original_mime_type.split('/')[-1]
-            if extension == 'jpeg':
-                extension = 'jpg'
-            suggested_filename = f"image_{message.id}.{extension}"
-        elif original_mime_type.startswith('video/'):
-            extension = original_mime_type.split('/')[-1]
-            suggested_filename = f"video_{message.id}.{extension}"
-        elif original_mime_type.startswith('audio/'):
-            extension = original_mime_type.split('/')[-1]
-            suggested_filename = f"audio_{message.id}.{extension}"
-```
-
-#### 3. Передача данных с именем файла
-```python
-# КЛЮЧЕВОЕ ИСПРАВЛЕНИЕ: Создаем объект с атрибутами для правильного отображения
-if downloaded_file and suggested_filename:
-    # Используем кортеж (data, filename) для передачи имени файла
-    media_files.append((downloaded_file, suggested_filename))
-else:
-    # Если не удалось определить имя файла, используем оригинальное медиа
-    media_files.append(message.media)
-```
-
-## 🧪 Тестирование
-
-Создан тест `test_album_fix_simple.py` для проверки логики:
-
-```bash
-python3 test_album_fix_simple.py
-```
-
-Результат теста:
-```
-✅ Тест логики извлечения имен файлов ПРОШЕЛ
-✅ Тест структуры кортежей ПРОШЕЛ  
-🎉 ВСЕ УПРОЩЕННЫЕ ТЕСТЫ ПРОШЛИ УСПЕШНО!
-```
-
-## 📋 Затронутые файлы
-
-### `copier.py`
-- **Метод `copy_album()`** - основное исправление для альбомов
-- **Метод `copy_single_message()`** - исправление для одиночных медиа файлов
-
-### Добавленные файлы
-- `test_album_fix.py` - полный тест с моками Telethon
-- `test_album_fix_simple.py` - упрощенный тест логики
-- `ALBUM_FIX_COMMENTS.md` - данная документация
-
-## 🔧 Техническая схема исправления
-
-### Было:
-```
-Message.media → download_media(file=bytes) → bytes object → send_file(bytes)
-                     ↓
-                Потеря метаданных: имя, расширение, MIME-тип
-                     ↓
-            Telegram сохраняет как "unnamed" файл
-```
-
-### Стало:
-```
-Message.media → извлечение атрибутов → download_media(file=bytes) → (bytes, filename) → send_file((bytes, filename))
-                     ↓                        ↓                           ↓
-              Сохранение метаданных    Скачивание данных        Передача с именем файла
-                     ↓                        ↓                           ↓
-              Генерация имени файла     bytes object           Telegram сохраняет с правильным именем
-```
-
-## ✨ Результат
-
-### До исправления:
-- 📁 `unnamed` (нет расширения, битый файл)
-- 📁 `unnamed` (нет расширения, битый файл)  
-- 📁 `unnamed` (нет расширения, битый файл)
-
-### После исправления:
-- 🖼️ `image_101.jpg` (правильное отображение)
-- 🎬 `video_102.mp4` (правильное отображение)
-- 🎵 `audio_103.mp3` (правильное отображение)
-
-## 🎯 Заключение
-
-**Критическая проблема с альбомами в комментариях полностью решена.**
-
-Теперь скрипт:
-- ✅ Правильно копирует текст комментариев
-- ✅ Сохраняет хронологический порядок
-- ✅ **Корректно отправляет альбомы с правильными именами файлов и расширениями**
-- ✅ Обеспечивает правильное отображение медиа в Telegram
-
-Исправление обратно совместимо и не влияет на копирование основных постов.
