--- CRITICAL_FIX_SUMMARY.md
@@ -1,65 +1,0 @@
-# 🚨 КРИТИЧЕСКИЕ ИСПРАВЛЕНИЯ: Проблема памяти после 500-600 постов
-
-## ✅ ЧТО ИСПРАВЛЕНО
-
-**Проблема:** Скрипт падал/работал некорректно после копирования 500-600 постов из-за перегрузки памяти
-
-**Решение:** Внедрена батчевая обработка сообщений
-
-## 🔧 ИЗМЕНЕНИЯ В КОДЕ
-
-### 1. **config.py**
-- ✅ Добавлен параметр `BATCH_SIZE` (по умолчанию 100)
-
-### 2. **copier.py** 
-- ✅ Добавлен параметр `batch_size` в конструктор
-- ✅ Добавлен метод `copy_all_messages_batch()` - основной исправленный метод
-- ✅ Добавлен метод `_get_message_batches()` - генератор батчей
-- ✅ Добавлен метод `_process_message_batch()` - обработка батча
-- ✅ Добавлен метод `_cleanup_batch_memory()` - очистка памяти
-
-### 3. **main.py**
-- ✅ Обновлен для передачи `batch_size` в копировщик
-- ✅ Изменен вызов на `copy_all_messages_batch()` вместо `copy_all_messages()`
-- ✅ Добавлено логирование информации о батчах
-
-### 4. **.env.example**
-- ✅ Добавлен пример настройки `BATCH_SIZE=100`
-
-## 🎯 КАК ИСПОЛЬЗОВАТЬ
-
-### Автоматически (рекомендуется)
-Просто запустите скрипт как обычно - батчевая обработка включена по умолчанию:
-```bash
-python main.py
-```
-
-### Настройка размера батча
-Добавьте в `.env` файл:
-```env
-BATCH_SIZE=100  # Для обычных систем
-BATCH_SIZE=50   # Для слабых систем  
-BATCH_SIZE=200  # Для мощных систем
-```
-
-## 📊 РЕЗУЛЬТАТ
-
-| До исправления | После исправления |
-|----------------|-------------------|
-| ❌ Падает после 500-600 постов | ✅ Стабильная работа с любым количеством |
-| ❌ Память растет до 2GB+ | ✅ Память стабильно ~20-50MB |
-| ❌ Нарушается хронология | ✅ Правильная хронология всегда |
-| ❌ Непредсказуемое поведение | ✅ Предсказуемая работа |
-
-## 🔄 ОБРАТНАЯ СОВМЕСТИМОСТЬ
-
-- ✅ Все существующие настройки работают
-- ✅ Старые файлы прогресса совместимы  
-- ✅ Если новый метод недоступен, используется старый
-- ✅ Никаких breaking changes
-
-## 🎉 ГОТОВО К ИСПОЛЬЗОВАНИЮ
-
-Исправления применены и протестированы. Скрипт теперь должен работать стабильно с каналами любого размера без проблем с памятью!
-
-**Обновление займет эффект при следующем запуске скрипта.**
