--- MEMORY_OPTIMIZATION.md
@@ -1,208 +1,0 @@
-# 🧠 Оптимизация памяти для TelegramCopier
-
-## 📋 Обзор
-
-Новая система оптимизации памяти решает проблемы с утечками памяти при работе с большими каналами, особенно с большим количеством комментариев. 
-
-### ⚡ Ключевые улучшения:
-
-- **🔻 Снижение использования памяти в 5-10 раз** (с 200-500 MB до 50-100 MB)
-- **📱 Поддержка слабых машин** с ограниченной памятью
-- **⚡ Минимальная потеря производительности** (5-15%)
-- **🛡️ Надежное восстановление** при сбоях
-- **📊 Мониторинг использования памяти** в реальном времени
-
-## 🔧 Архитектура
-
-### Двухуровневая система кэширования:
-
-1. **Level 1: LRU кэш в памяти** 
-   - Хранит 1000 наиболее часто используемых постов с комментариями
-   - Автоматическое вытеснение редко используемых данных
-
-2. **Level 2: SQLite база данных**
-   - Постоянное хранение всех комментариев на диске
-   - Быстрое извлечение по индексированным ключам
-   - Автоматическая очистка старых данных
-
-### Контроль памяти:
-
-- **Автоматический мониторинг** использования памяти через psutil
-- **Принудительная очистка** при превышении лимитов
-- **Адаптивная батчевая обработка** в зависимости от доступной памяти
-
-## ⚙️ Настройка
-
-### В .env файле:
-
-```bash
-# Настройки оптимизации памяти
-ENABLE_MEMORY_OPTIMIZATION=true     # Включить оптимизацию
-MEMORY_LIMIT_MB=100                 # Лимит памяти (для слабых машин: 50-100)
-LRU_CACHE_SIZE=1000                 # Размер кэша в памяти
-```
-
-### Рекомендации по настройке:
-
-| Тип машины | MEMORY_LIMIT_MB | LRU_CACHE_SIZE | Ожидаемая производительность |
-|------------|-----------------|----------------|-------------------------------|
-| **Слабая** (2-4 GB RAM) | 50 | 500 | 85-90% от исходной |
-| **Средняя** (4-8 GB RAM) | 100 | 1000 | 90-95% от исходной |
-| **Мощная** (8+ GB RAM) | 200 | 2000 | 95-98% от исходной |
-
-## 📊 Сравнение производительности
-
-### До оптимизации:
-```
-🔴 Проблемы:
-- Память: 200-500 MB для больших каналов
-- Сканирование 5000+ сообщений для поиска discussion groups
-- Загрузка ВСЕХ комментариев в память одновременно
-- Утечки памяти при длительной работе
-- Сбои на слабых машинах
-```
-
-### После оптимизации:
-```
-✅ Улучшения:
-- Память: 50-100 MB (снижение в 5-10 раз)
-- Адаптивное сканирование (1000-2000 сообщений)
-- Батчевая загрузка комментариев по мере необходимости
-- Автоматическая очистка памяти
-- Стабильная работа на любых машинах
-```
-
-## 🚀 Использование
-
-### Автоматическое включение
-
-Оптимизация включается автоматически при создании TelegramCopier:
-
-```python
-# В main.py автоматически применяется:
-if config.enable_memory_optimization:
-    apply_memory_optimization(copier, config.memory_limit_mb)
-```
-
-### Ручное управление
-
-```python
-from memory_optimization_patch import apply_memory_optimization
-
-# Применить оптимизацию к существующему копировщику
-apply_memory_optimization(copier, memory_limit_mb=100)
-
-# Получить статистику
-stats = copier.get_memory_optimization_stats()
-print(f"Cache hit ratio: {stats['cache_hit_ratio']:.1%}")
-print(f"Memory usage: {stats['total_memory_mb']:.1f} MB")
-```
-
-## 📈 Мониторинг
-
-### Логи производительности:
-
-```
-🧠 Применяем оптимизацию памяти (лимит: 100 MB)
-📊 ОПТИМИЗИРОВАННАЯ ПРЕДЗАГРУЗКА: Сканируем канал...
-💾 Сохранено 1543 связей пост->комментарии в базу данных
-💨 Кэш-попадание для поста 12345: 15 комментариев
-🧹 Очистка памяти: освобожден кэш (1000 элементов)
-```
-
-### Финальная статистика:
-
-```
-📊 ФИНАЛЬНАЯ СТАТИСТИКА ОПТИМИЗАЦИИ ПАМЯТИ:
-   💨 Cache hit ratio: 89.2%
-   💾 Использование памяти: 87.3 MB
-   🗄️ Размер базы данных: 12.5 MB
-   🧹 Очисток памяти: 3
-```
-
-## 🔍 Техническая реализация
-
-### Ключевые компоненты:
-
-1. **OptimizedCommentsStorage** (`optimized_cache.py`)
-   - Управление SQLite базой данных
-   - LRU кэш с автоматическим вытеснением
-   - Мониторинг памяти через psutil
-
-2. **Memory Optimization Patch** (`memory_optimization_patch.py`)
-   - Интеграционный слой для существующего кода
-   - Замена методов кэширования без изменения API
-   - Автоматическая миграция данных
-
-3. **Enhanced Configuration** (`config.py`)
-   - Новые настройки оптимизации памяти
-   - Валидация лимитов памяти
-   - Адаптивные значения по умолчанию
-
-### Алгоритм работы:
-
-1. **Инициализация**: Создание SQLite базы и LRU кэша
-2. **Предзагрузка**: Адаптивное сканирование канала для поиска discussion groups
-3. **Батчевая обработка**: Загрузка комментариев батчами в базу данных
-4. **Кэширование**: Автоматическое управление памятью через LRU
-5. **Мониторинг**: Постоянный контроль использования памяти
-6. **Очистка**: Автоматическое освобождение памяти при превышении лимитов
-
-## 🛠️ Совместимость
-
-### Обратная совместимость:
-- ✅ Полная совместимость с существующим API
-- ✅ Автоматическая миграция при первом запуске  
-- ✅ Возможность отключения оптимизации (`ENABLE_MEMORY_OPTIMIZATION=false`)
-
-### Зависимости:
-- `psutil>=5.9.0` - для мониторинга памяти
-- `sqlite3` - встроен в Python (база данных)
-
-## 🔧 Устранение неполадок
-
-### Частые проблемы:
-
-**Q: Высокое использование памяти несмотря на оптимизацию**
-```bash
-# Уменьшите лимиты в .env:
-MEMORY_LIMIT_MB=50
-LRU_CACHE_SIZE=500
-```
-
-**Q: Медленная работа после включения оптимизации**
-```bash
-# Увеличьте размер кэша:
-LRU_CACHE_SIZE=2000
-# Или увеличьте лимит памяти:
-MEMORY_LIMIT_MB=200
-```
-
-**Q: Ошибки базы данных**
-```bash
-# Удалите файл базы данных для пересоздания:
-rm comments_cache.db
-```
-
-### Отладка:
-
-```python
-# Включите детальные логи
-import logging
-logging.getLogger('telegram_copier.optimized_cache').setLevel(logging.DEBUG)
-
-# Получите статистику производительности
-stats = copier.get_memory_optimization_stats()
-for key, value in stats.items():
-    print(f"{key}: {value}")
-```
-
-## 📝 Changelog
-
-### v1.0.0 (2024-08-15)
-- ✅ Двухуровневая система кэширования (LRU + SQLite)
-- ✅ Автоматический мониторинг памяти
-- ✅ Адаптивная батчевая обработка  
-- ✅ Полная интеграция с существующим кодом
-- ✅ Настройки конфигурации для разных типов машин
-- ✅ Статистика производительности в реальном времени
