--- QUICK_FIX_GUIDE.md
@@ -1,75 +1,0 @@
-# 🚀 БЫСТРОЕ ИСПРАВЛЕНИЕ: Батчевая обработка
-
-## 🎯 КРАТКО О ПРОБЛЕМЕ
-Скрипт падает после 500-600 постов из-за нехватки памяти. **Решение: батчевая обработка по 100 сообщений**.
-
-## ⚡ БЫСТРОЕ ИСПРАВЛЕНИЕ
-
-### Вариант 1: Добавить параметр в конфигурацию
-
-**1. Добавьте в `.env` файл:**
-```env
-BATCH_SIZE=100
-```
-
-**2. Добавьте в `config.py` строку 39:**
-```python
-self.batch_size: int = int(os.getenv('BATCH_SIZE', '100'))
-```
-
-**3. Замените в `main.py` строку ~343:**
-```python
-# БЫЛО:
-self.copier = TelegramCopier(...)
-
-# СТАЛО:
-from copier_fixed import TelegramCopierFixed
-self.copier = TelegramCopierFixed(..., batch_size=self.config.batch_size)
-```
-
-**4. Замените в `main.py` строку ~364:**
-```python
-# БЫЛО:
-stats = await self.copier.copy_all_messages(resume_from_id)
-
-# СТАЛО:
-stats = await self.copier.copy_all_messages_batch(resume_from_id)
-```
-
-### Вариант 2: Простая замена файла
-
-**1. Переименуйте файлы:**
-```bash
-mv copier.py copier_original.py
-mv copier_fixed.py copier.py
-```
-
-**2. Скопируйте недостающие методы из `copier_original.py` в `copier.py`**
-
-**3. Запустите как обычно:**
-```bash
-python main.py
-```
-
-## 🔧 НАСТРОЙКИ БАТЧЕЙ
-
-| Система | Batch Size | Память |
-|---------|------------|--------|
-| Слабая | 25-50 | <512MB |
-| Обычная | 100 | <1GB |
-| Мощная | 150-200 | <2GB |
-
-## 🎉 РЕЗУЛЬТАТ
-
-- ✅ Стабильная работа с любым количеством постов
-- ✅ Память всегда <50MB 
-- ✅ Правильная хронология
-- ✅ Возобновление работы с любого места
-
-## 🆘 ЕСЛИ ПРОБЛЕМЫ ОСТАЛИСЬ
-
-1. Уменьшите `BATCH_SIZE=25`
-2. Проверьте свободную память: `free -h`
-3. Перезапустите скрипт с места остановки
-
-**Исправление решает проблему в 99% случаев!**
