--- README_V3.md
@@ -1,141 +1,0 @@
-# 🚀 Telegram Copier v3.0 - ПОЛНОСТЬЮ ИСПРАВЛЕНО
-
-## ✅ Все критические проблемы РЕШЕНЫ
-
-### Старые проблемы (v2.x):
-❌ Комментарии не копируются как нужно  
-❌ Альбомы разваливаются на отдельные сообщения  
-❌ Ошибки "Cannot use None as file" и MediaProxy  
-❌ Нет сохранения данных при перезапуске  
-❌ Неправильный хронологический порядок  
-
-### Новые решения (v3.0):
-✅ **Комментарии копируются правильно** - прямое использование Telethon API  
-✅ **Альбомы отправляются целиком** - группировка по `grouped_id`  
-✅ **Никаких MediaProxy ошибок** - прямое получение оригинальных сообщений  
-✅ **Персистентная SQLite БД** - данные сохраняются между перезапусками  
-✅ **Правильный хронологический порядок** - сначала посты, потом комментарии  
-
-## 🎯 Как это должно работать (и теперь работает):
-
-```
-1. 📝 Копируем пост №1: "Доброе утро!" (10:00)
-   💬 Копируем комментарий (10:01)
-   💬 Копируем комментарий (10:02)
-
-2. 📷 Копируем альбом №2 ЦЕЛИКОМ (10:05)
-   [отправляем 3 фото одновременно как альбом]
-   💬 Копируем комментарий (10:06)
-   💬 Копируем комментарий (10:07)
-```
-
-## 🚀 Запуск
-
-### Новая версия (РЕКОМЕНДУЕТСЯ):
-```bash
-python main_v3.py
-```
-
-### Старая версия (НЕ РЕКОМЕНДУЕТСЯ):
-```bash
-python main.py  # Содержит все старые проблемы
-```
-
-## 🔧 Настройка
-
-Файл `.env` остается тем же:
-```env
-SOURCE_GROUP_ID=your_source_channel
-TARGET_GROUP_ID=your_target_channel
-API_ID=your_api_id
-API_HASH=your_api_hash
-PHONE=your_phone
-```
-
-## 📊 Архитектура v3.0
-
-### 1. **Этап сканирования** (выполняется один раз):
-- Полное сканирование исходного канала
-- Сохранение всех постов в SQLite таблицу `posts`
-- Поиск и сохранение всех комментариев в таблицу `comments`
-- Автоматическое определение discussion groups
-
-### 2. **Этап копирования** (в хронологическом порядке):
-- Загрузка данных из БД (мгновенно)
-- Группировка альбомов по `grouped_id`
-- Копирование постов в хронологическом порядке
-- Копирование комментариев в хронологическом порядке
-
-## 🗄️ База данных
-
-```sql
--- Посты канала
-CREATE TABLE posts (
-    id INTEGER PRIMARY KEY,
-    message_text TEXT,
-    grouped_id INTEGER,  -- Для альбомов
-    processed BOOLEAN DEFAULT FALSE
-);
-
--- Комментарии к постам
-CREATE TABLE comments (
-    id INTEGER PRIMARY KEY,
-    post_id INTEGER,     -- Связь с постом
-    processed BOOLEAN DEFAULT FALSE,
-    FOREIGN KEY (post_id) REFERENCES posts (id)
-);
-```
-
-## 🐛 Исправленные ошибки
-
-### Старые логи (ПРОБЛЕМНЫЕ):
-```
-WARNING - Альбом не содержит медиа файлов
-WARNING - Сообщение 9825 является частью альбома, но обрабатывается как одиночное  
-WARNING - Не удалось отправить медиа MediaProxy: Cannot use None as file
-```
-
-### Новые логи (ИСПРАВЛЕННЫЕ):
-```
-✅ Альбом 13961310187141506 (3 сообщения) скопирован
-✅ Пост 9825 -> 12345
-💬 Комментарий 9826 -> 12346 (reply to: 12345)
-📊 Статистика: Посты: 1450/1500, Альбомы: 45/50, Комментарии: 250/300
-```
-
-## 📈 Производительность
-
-| Операция | v2.x | v3.0 |
-|----------|------|------|
-| Перезапуск | Полное пересканирование | Загрузка из БД (мгновенно) |
-| Альбомы | Отдельные сообщения | Целые альбомы |
-| Комментарии | Не работает стабильно | Работает всегда |
-| Порядок | Случайный | Строго хронологический |
-| Данные | Теряются при перезапуске | Сохраняются в БД |
-
-## 🔍 Мониторинг
-
-### Проверка прогресса:
-```bash
-# Сколько постов обработано
-sqlite3 telegram_copier_v3.db "SELECT COUNT(*) FROM posts WHERE processed = 1;"
-
-# Сколько комментариев обработано  
-sqlite3 telegram_copier_v3.db "SELECT COUNT(*) FROM comments WHERE processed = 1;"
-
-# Статистика альбомов
-sqlite3 telegram_copier_v3.db "SELECT grouped_id, COUNT(*) FROM posts WHERE grouped_id IS NOT NULL GROUP BY grouped_id;"
-```
-
-## 🎉 Результат
-
-**v3.0 решает ВСЕ критические проблемы предыдущих версий:**
-
-✅ Комментарии копируются как задумано  
-✅ Альбомы отправляются целиком с правильным текстом  
-✅ Правильный хронологический порядок  
-✅ Персистентное хранение - нет потери данных  
-✅ Поддержка всех типов медиа в постах и комментариях  
-✅ Никаких ошибок MediaProxy  
-
-**Просто запустите `python main_v3.py` и наслаждайтесь стабильной работой!**
