--- copied_messages.json
@@ -1,10 +1,0 @@
-{
-  "copied_messages": {},
-  "statistics": {
-    "total_copied": 0,
-    "total_failed": 0,
-    "last_updated": "2025-08-21T23:35:03.227276"
-  },
-  "source_channel": "-1002399927446",
-  "target_channel": "-1002724748282"
-}
