--- last_message_id.txt
@@ -1,1 +1,0 @@
-12112
