--- test_anti_nesting.py
@@ -1,144 +1,0 @@
-#!/usr/bin/env python3
-"""
-Демонстрация режима антивложенности - превращение вложенной структуры в плоскую.
-"""
-
-import asyncio
-import logging
-
-# Настройка логирования для теста
-logging.basicConfig(
-    level=logging.INFO,
-    format='%(asctime)s - %(levelname)s - %(message)s'
-)
-
-class AntiNestingDemo:
-    """Демонстрация режима антивложенности."""
-    
-    def __init__(self):
-        self.logger = logging.getLogger('anti_nesting_demo')
-    
-    def demonstrate_concepts(self):
-        """Демонстрирует концепцию антивложенности."""
-        
-        self.logger.info("=" * 80)
-        self.logger.info("🔄 ДЕМОНСТРАЦИЯ РЕЖИМА АНТИВЛОЖЕННОСТИ")
-        self.logger.info("=" * 80)
-        
-        # Исходная структура
-        self.logger.info("\n📋 ИСХОДНАЯ СТРУКТУРА В ТЕЛЕГРАМ КАНАЛЕ:")
-        original_structure = [
-            "📌 Пост 1: 'Привет всем!' (ID: 100)",
-            "  ├─ 💬 Комментарий: 'Привет и тебе!' (ID: 101)", 
-            "  └─ 💬 Комментарий: 'Как дела?' (ID: 102)",
-            "",
-            "🎬 Альбом: Фотографии отпуска (ID: 103-105)",
-            "  ├─ 💬 Комментарий: 'Красивые места!' (ID: 106)",
-            "  ├─ 🎬 Альбом в комментарии: еще фото (ID: 107-108)",
-            "  └─ 💬 Комментарий: 'Где это было?' (ID: 109)",
-            "",
-            "📌 Пост 2: 'Важная новость' (ID: 110)",
-            "  └─ 💬 Комментарий: 'Спасибо за информацию' (ID: 111)",
-            "      └─ 💬 Ответ: 'Пожалуйста!' (ID: 112)"
-        ]
-        
-        for line in original_structure:
-            self.logger.info(f"   {line}")
-        
-        # Режим с вложенностью
-        self.logger.info("\n✅ РЕЖИМ С СОХРАНЕНИЕМ ВЛОЖЕННОСТИ (flatten_structure=False):")
-        nested_result = [
-            "📌 Пост 1 (ID: 100)",
-            "💬 Комментарий (ID: 101)", 
-            "💬 Комментарий (ID: 102)",
-            "🎬 Альбом (ID: 103-105)",
-            "💬 Комментарий к альбому (ID: 106)",
-            "🎬 Альбом в комментарии (ID: 107-108)", 
-            "💬 Комментарий к альбому (ID: 109)",
-            "📌 Пост 2 (ID: 110)",
-            "💬 Комментарий (ID: 111)",
-            "💬 Ответ на комментарий (ID: 112)"
-        ]
-        
-        for i, line in enumerate(nested_result, 1):
-            self.logger.info(f"   {i:2d}. {line}")
-        
-        self.logger.info("   ➤ Структура сохранена: комментарии следуют сразу за постами")
-        
-        # Режим антивложенности
-        self.logger.info("\n🔄 РЕЖИМ АНТИВЛОЖЕННОСТИ (flatten_structure=True):")
-        flattened_result = [
-            "📌 Пост 1 (ID: 100)",
-            "💬 Комментарий (ID: 101) ← превращен в обычный пост", 
-            "💬 Комментарий (ID: 102) ← превращен в обычный пост",
-            "🎬 Альбом (ID: 103-105)",
-            "💬 Комментарий (ID: 106) ← превращен в обычный пост",
-            "🎬 Альбом (ID: 107-108) ← был в комментарии, стал обычным альбомом", 
-            "💬 Комментарий (ID: 109) ← превращен в обычный пост",
-            "📌 Пост 2 (ID: 110)",
-            "💬 Комментарий (ID: 111) ← превращен в обычный пост",
-            "💬 Ответ (ID: 112) ← превращен в обычный пост"
-        ]
-        
-        for i, line in enumerate(flattened_result, 1):
-            self.logger.info(f"   {i:2d}. {line}")
-        
-        self.logger.info("   ➤ Все в хронологическом порядке, но без вложенности!")
-        
-        # Ключевые отличия
-        self.logger.info("\n🔍 КЛЮЧЕВЫЕ ОТЛИЧИЯ:")
-        differences = [
-            "🎯 Хронология: Строго по ID сообщений (100, 101, 102, 103-105, 106, 107-108, 109, 110, 111, 112)",
-            "📱 Комментарии: Становятся обычными постами без связи с родительскими",
-            "🎬 Альбомы: Остаются альбомами, но теряют контекст 'из комментария'",
-            "📊 Структура: Плоская последовательность вместо дерева",
-            "⚡ Простота: Легче для восприятия как обычная лента постов"
-        ]
-        
-        for diff in differences:
-            self.logger.info(f"   • {diff}")
-        
-        # Настройка
-        self.logger.info("\n⚙️ КАК ВКЛЮЧИТЬ РЕЖИМ АНТИВЛОЖЕННОСТИ:")
-        setup_instructions = [
-            "1. В файле .env добавить: FLATTEN_STRUCTURE=true",
-            "2. Или в переменной окружения: export FLATTEN_STRUCTURE=true", 
-            "3. Запустить копировщик как обычно: python main.py",
-            "4. Все комментарии автоматически станут обычными постами"
-        ]
-        
-        for instruction in setup_instructions:
-            self.logger.info(f"   {instruction}")
-        
-        # Сценарии использования
-        self.logger.info("\n💡 КОГДА ИСПОЛЬЗОВАТЬ КАЖДЫЙ РЕЖИМ:")
-        use_cases = [
-            "",
-            "🔗 ВЛОЖЕННЫЙ РЕЖИМ (flatten_structure=False) - для:",
-            "   • Полного сохранения структуры обсуждений",
-            "   • Архивирования форумов и групп с активными обсуждениями", 
-            "   • Сохранения контекста 'что к чему относится'",
-            "",
-            "📋 ПЛОСКИЙ РЕЖИМ (flatten_structure=True) - для:",
-            "   • Создания хронологической ленты новостей",
-            "   • Копирования в каналы без комментариев",
-            "   • Упрощения структуры для лучшего восприятия",
-            "   • Избежания 'вложенного хаоса' в больших группах"
-        ]
-        
-        for case in use_cases:
-            self.logger.info(f"   {case}")
-        
-        self.logger.info("\n" + "=" * 80)
-        self.logger.info("✅ Оба режима сохраняют правильную логику альбомов!")
-        self.logger.info("✅ Оба режима соблюдают строгую хронологию!")
-        self.logger.info("✅ Переключение между режимами - одна строчка в конфиге!")
-        self.logger.info("=" * 80)
-
-def main():
-    """Запуск демонстрации."""
-    demo = AntiNestingDemo()
-    demo.demonstrate_concepts()
-
-if __name__ == "__main__":
-    main()
