--- test_minimal.py
@@ -1,252 +1,0 @@
-#!/usr/bin/env python3
-"""
-Минимальные тесты без внешних зависимостей.
-"""
-
-import os
-import sys
-import tempfile
-
-
-def test_progress_saving_logic():
-    """Тест логики сохранения прогресса (без импорта utils)."""
-    print("Тестируем логику сохранения прогресса...")
-    
-    def save_progress(message_id: int, filename: str) -> None:
-        """Упрощенная версия save_last_message_id для тестирования."""
-        try:
-            # Атомарная запись через временный файл
-            temp_filename = f"{filename}.tmp"
-            with open(temp_filename, 'w', encoding='utf-8') as f:
-                f.write(str(message_id))
-                f.flush()
-                os.fsync(f.fileno())
-            
-            # Атомарное переименование
-            if os.path.exists(filename):
-                os.replace(temp_filename, filename)
-            else:
-                os.rename(temp_filename, filename)
-        except Exception as e:
-            print(f"Ошибка сохранения: {e}")
-            if os.path.exists(f"{filename}.tmp"):
-                os.remove(f"{filename}.tmp")
-    
-    def load_progress(filename: str) -> int:
-        """Упрощенная версия load_last_message_id для тестирования."""
-        try:
-            if os.path.exists(filename):
-                with open(filename, 'r', encoding='utf-8') as f:
-                    return int(f.read().strip())
-        except Exception as e:
-            print(f"Ошибка загрузки: {e}")
-        return None
-    
-    with tempfile.TemporaryDirectory() as temp_dir:
-        progress_file = os.path.join(temp_dir, 'progress.txt')
-        
-        # Тест сохранения
-        save_progress(12345, progress_file)
-        assert os.path.exists(progress_file), "Файл прогресса не создан"
-        
-        # Тест загрузки
-        loaded_id = load_progress(progress_file)
-        assert loaded_id == 12345, f"Ожидался ID 12345, получен {loaded_id}"
-        
-        # Тест что временный файл удален
-        temp_file = f"{progress_file}.tmp"
-        assert not os.path.exists(temp_file), "Временный файл не удален"
-        
-        # Тест перезаписи
-        save_progress(67890, progress_file)
-        loaded_id = load_progress(progress_file)
-        assert loaded_id == 67890, f"Ожидался ID 67890, получен {loaded_id}"
-        
-        print("✅ Логика атомарного сохранения прогресса корректна")
-
-
-def test_file_lock_logic():
-    """Тест логики блокировки файлов (упрощенная версия)."""
-    print("Тестируем логику блокировки файлов...")
-    
-    import fcntl
-    
-    with tempfile.TemporaryDirectory() as temp_dir:
-        lock_file = os.path.join(temp_dir, 'test.lock')
-        
-        # Первая блокировка
-        fd1 = open(lock_file, 'w')
-        try:
-            fcntl.flock(fd1.fileno(), fcntl.LOCK_EX | fcntl.LOCK_NB)
-            fd1.write(str(os.getpid()))
-            fd1.flush()
-            
-            # Попытка второй блокировки должна провалиться
-            fd2 = open(lock_file, 'w')
-            try:
-                fcntl.flock(fd2.fileno(), fcntl.LOCK_EX | fcntl.LOCK_NB)
-                fd2.close()
-                assert False, "Вторая блокировка не должна была быть получена"
-            except (IOError, OSError):
-                # Ожидаемое поведение - блокировка не получена
-                fd2.close()
-                pass
-            
-            # Освобождаем первую блокировку
-            fcntl.flock(fd1.fileno(), fcntl.LOCK_UN)
-            fd1.close()
-            
-            # Теперь новая блокировка должна работать
-            fd3 = open(lock_file, 'w')
-            try:
-                fcntl.flock(fd3.fileno(), fcntl.LOCK_EX | fcntl.LOCK_NB)
-                fcntl.flock(fd3.fileno(), fcntl.LOCK_UN)
-                fd3.close()
-            except (IOError, OSError):
-                fd3.close()
-                assert False, "Блокировка после освобождения должна работать"
-            
-        except (IOError, OSError):
-            fd1.close()
-            assert False, "Первая блокировка должна была быть получена"
-    
-    print("✅ Логика блокировки файлов корректна")
-
-
-def test_config_logic():
-    """Тест логики конфигурации."""
-    print("Тестируем логику конфигурации...")
-    
-    # Имитируем класс Config без внешних зависимостей
-    class MockConfig:
-        def __init__(self):
-            self.api_id = int(os.getenv('API_ID', '0'))
-            self.api_hash = os.getenv('API_HASH', '')
-            self.phone = os.getenv('PHONE', '')
-            self.source_group_id = os.getenv('SOURCE_GROUP_ID', '')
-            self.target_group_id = os.getenv('TARGET_GROUP_ID', '')
-        
-        def validate(self) -> bool:
-            required_fields = [
-                ('API_ID', self.api_id),
-                ('API_HASH', self.api_hash),
-                ('PHONE', self.phone),
-                ('SOURCE_GROUP_ID', self.source_group_id),
-                ('TARGET_GROUP_ID', self.target_group_id)
-            ]
-            
-            for field_name, field_value in required_fields:
-                if not field_value:
-                    return False
-            
-            if self.api_id == 0:
-                return False
-            
-            return True
-    
-    # Тест с пустыми переменными
-    for key in ['API_ID', 'API_HASH', 'PHONE', 'SOURCE_GROUP_ID', 'TARGET_GROUP_ID']:
-        if key in os.environ:
-            del os.environ[key]
-    
-    config = MockConfig()
-    assert not config.validate(), "Пустая конфигурация должна быть невалидной"
-    
-    # Тест с валидными данными
-    os.environ['API_ID'] = '12345678'
-    os.environ['API_HASH'] = 'test_hash'
-    os.environ['PHONE'] = '+1234567890'
-    os.environ['SOURCE_GROUP_ID'] = '@source'
-    os.environ['TARGET_GROUP_ID'] = '@target'
-    
-    config = MockConfig()
-    assert config.validate(), "Валидная конфигурация должна проходить проверку"
-    
-    # Очищаем переменные
-    for key in ['API_ID', 'API_HASH', 'PHONE', 'SOURCE_GROUP_ID', 'TARGET_GROUP_ID']:
-        if key in os.environ:
-            del os.environ[key]
-    
-    print("✅ Логика конфигурации корректна")
-
-
-def test_utility_functions():
-    """Тест вспомогательных функций."""
-    print("Тестируем вспомогательные функции...")
-    
-    def format_file_size(size_bytes: int) -> str:
-        """Локальная версия функции для тестирования."""
-        if size_bytes == 0:
-            return "0 B"
-        
-        size_names = ["B", "KB", "MB", "GB"]
-        i = 0
-        size = float(size_bytes)
-        
-        while size >= 1024.0 and i < len(size_names) - 1:
-            size /= 1024.0
-            i += 1
-        
-        return f"{size:.1f} {size_names[i]}"
-    
-    def sanitize_filename(filename: str) -> str:
-        """Локальная версия функции для тестирования."""
-        invalid_chars = '<>:"/\\|?*'
-        for char in invalid_chars:
-            filename = filename.replace(char, '_')
-        
-        if len(filename) > 255:
-            name, ext = os.path.splitext(filename)
-            filename = name[:255-len(ext)] + ext
-        
-        return filename
-    
-    # Тесты форматирования размера
-    assert format_file_size(0) == "0 B"
-    assert format_file_size(1024) == "1.0 KB"
-    assert format_file_size(1024 * 1024) == "1.0 MB"
-    assert format_file_size(1024 * 1024 * 1024) == "1.0 GB"
-    
-    # Тесты очистки имени файла
-    assert sanitize_filename("normal.txt") == "normal.txt"
-    assert sanitize_filename("file<>test.txt") == "file__test.txt"
-    assert sanitize_filename('file"test|file.txt') == "file_test_file.txt"
-    
-    # Тест длинного имени
-    long_name = "a" * 300 + ".txt"
-    sanitized = sanitize_filename(long_name)
-    assert len(sanitized) <= 255, f"Имя слишком длинное: {len(sanitized)}"
-    
-    print("✅ Вспомогательные функции работают корректно")
-
-
-def main():
-    """Запуск всех минимальных тестов."""
-    print("=== Запуск минимальных тестов логики ===")
-    
-    try:
-        test_progress_saving_logic()
-        test_file_lock_logic()
-        test_config_logic()
-        test_utility_functions()
-        
-        print("\n🎉 Все минимальные тесты пройдены!")
-        print("\n✅ ПРОВЕРЕНО:")
-        print("  - Атомарное сохранение прогресса")
-        print("  - Блокировка файлов работает")
-        print("  - Логика валидации конфигурации")
-        print("  - Вспомогательные функции")
-        print("  - Обработка граничных случаев")
-        
-        return 0
-        
-    except Exception as e:
-        print(f"\n❌ Ошибка в тестах: {e}")
-        import traceback
-        traceback.print_exc()
-        return 1
-
-
-if __name__ == '__main__':
-    exit_code = main()
-    sys.exit(exit_code)
